Add files via upload
Rock, Paper, Scissors Game

Welcome to the classic game of Rock, Paper, Scissors brought to life on the web! This interactive game is developed using HTML, CSS, and JavaScript to provide a seamless and enjoyable user experience.

Features:

User-friendly Interface: The game boasts an intuitive and visually appealing interface, making it easy for players to engage and have fun.

Dynamic Gameplay: Users can make their choice of Rock, Paper, or Scissors, and the game dynamically responds, revealing the computer's random choice. The winner is determined by the timeless rules of the game.

Real-time Results: The outcome of each round is instantly displayed on the screen, allowing players to see who emerges victorious in the battle of elements.

Open Source: This project is open-source and hosted on GitHub, making it accessible for exploration, learning, and collaboration. Feel free to dive into the code, contribute enhancements, or use it as a foundation for your own projects.

How to Play:

Clone or download the repository to your local machine.
Open the index.html file in your web browser.
Make your choice by clicking on Rock, Paper, or Scissors.
Enjoy the game and see if you can outsmart the computer!
Whether you're a coding enthusiast looking to learn more about web development or just seeking a quick and entertaining game, this Rock, Paper, Scissors project is a perfect fit. Have fun playing and exploring the code!